requirements.txt
Flask==2.3.2
python-telegram-bot==13.15